add products curd apis

+++ src/main/java/com/svr/ecommerce/dtos/ProductDto.java
@@ -16,5 +16,5 @@ public class ProductDto {
     private String name;
     private String description;
     private BigDecimal price;
-    private Long categoryId;
+    private Byte categoryId;
 }
